pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "d3-dna-discrete-diffusion"
version = "0.1.0"
description = "D3: DNA Discrete Diffusion - A diffusion model for DNA sequence generation"
readme = "README.md"
authors = [
    {name = "D3 DNA Team"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision",
    "pytorch-lightning",
    "transformers",
    "hydra-core",
    "omegaconf",
    "wandb",
    "dm-tree",
    "biotite",
    "matplotlib",
    "seaborn",
    "h5py",
    "numpy",
    "scipy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*", "model*", "scripts*", "evaluation*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"] 